+++ src/application_settings/configuring_base.py
@@ -3,15 +3,24 @@ from typing import Any, TypeVar
 
 from pydantic.dataclasses import dataclass
 
-from .container_base import ContainerBase, ContainerSectionBase, ContainerTypeStr
+from application_settings.container_base import ContainerBase, FileFormat
+from application_settings.container_section_base import (
+    ContainerSectionBase,
+    SectionTypeStr,
+)
 
 ConfigT = TypeVar("ConfigT", bound="ConfigBase")
-ConfigSectionT = TypeVar("ConfigSectionT", bound="ConfigSectionBase")
+ConfigT.__doc__ = "Represents ConfigBase and all subclasses"
 
 
 @dataclass(frozen=True)
 class ConfigSectionBase(ContainerSectionBase):
     """Base class for all ConfigSection classes (so that we can bound a TypeVar)"""
+
+    @classmethod
+    def kind_string(cls) -> SectionTypeStr:
+        "Return 'Config'"
+        return "Config"
 
 
 @dataclass(frozen=True)
@@ -19,11 +28,19 @@ class ConfigBase(ContainerBase):
     """Base class for main Config class"""
 
     @classmethod
-    def kind_string(cls: type[ConfigT]) -> ContainerTypeStr:
+    def kind_string(cls) -> SectionTypeStr:
         "Return 'Config'"
         return "Config"
 
-    def update(self: ConfigT, changes: dict[str, dict[str, Any]]) -> ConfigT:
+    @classmethod
+    def default_file_format(cls) -> FileFormat:
+        "Return the default file format"
+        return FileFormat.TOML
+
+    @classmethod
+    def update(
+        cls: type["ConfigBase"], changes: dict[str, dict[str, Any]]
+    ) -> "ConfigBase":
         "Update and save the settings with data specified in changes; not meant for config"
         raise TypeError(
             "Configs should not be updated runtime; consider converting to settings."